Updated to long

+++ multiplier.c
@@ -8,7 +8,7 @@
 #include <stdio.h>
 #include <time.h>
 /* This function applies the algorithm resursively */
-int multiplierRecursive(int a, int b){
+long multiplierRecursive(long a, long b){
   if(a == 0){
     return 0;
   }
@@ -26,9 +26,9 @@ int multiplierRecursive(int a, int b){
 }
 
 /* This function applies the algorithm non resursively */
-int multiplier(int a, int b){
+long multiplier(long a, long b){
 
-  int c = 0;
+  long c = 0;
 
   while(a > 1){
     if(a%2 == 1){
@@ -43,29 +43,29 @@ int multiplier(int a, int b){
 
 /* Main function to run algorithm and get input */
 int main(){
-  int a, b, c; //To store values
+  long a, b, c; //To store values
   double time_spent; //To store time
   clock_t begin, end; //For timing
 
   //Get input from user
   printf("Please enter your first number: ");
-  scanf("%d", &a);
+  scanf("%ld", &a);
   printf("Please enter your second number: ");
-  scanf("%d", &b);
+  scanf("%ld", &b);
 
   //Call recursive function, time and print result
   begin = clock();
   c = multiplierRecursive(a, b);
   end = clock();
   time_spent = (double)(end - begin) / CLOCKS_PER_SEC;
-  printf("Recursively:%d * %d = %d\nTime elapsed: %.10lf seconds\n", a, b, c, time_spent);
+  printf("Recursively:%ld * %ld = %ld\nTime elapsed: %.10lf seconds\n", a, b, c, time_spent);
 
   //Call non-recursive function, time and print result
   begin = clock();
   c = multiplierRecursive(a, b);
   end = clock();
   time_spent = (double)(end - begin) / CLOCKS_PER_SEC;
-  printf("Non Recursively:%d * %d = %d\nTime elapsed: %.10lf seconds\n", a, b, c, time_spent);
+  printf("Non Recursively:%ld * %ld = %ld\nTime elapsed: %.10lf seconds\n", a, b, c, time_spent);
 
   return 1;
 }
